fix sync

+++ camera/src/main/java/com/underdog/raver/videoTrimmer.java
@@ -66,7 +66,7 @@ public class videoTrimmer extends AppCompatActivity {
 
     boolean isPlaying = false;
 
-    int duration;
+    double duration;
     String filePrefix;
     String filePrefix_change;
     String filePrefix_merge;
@@ -246,7 +246,7 @@ public class videoTrimmer extends AppCompatActivity {
         filePath_mp4_change = dest_change_mp4.getAbsolutePath();
         //Toast.makeText(this, filePath, Toast.LENGTH_SHORT).show();
 
-        command = new String[]{"-ss", "" + startMs / 1000, "-y", "-i", ori_Path, "-t", "" + (endMs - startMs) / 1000, "-vcodec", "mpeg4", "-b:v", "2097152", "-b:a", "48000", "-ac", "2", "-ar", "22050", filePath};
+        command = new String[]{"-ss", "" + startMs / 1000, "-y", "-i", ori_Path, "-t", "" + (endMs-startMs) / 1000,"-vcodec", "mpeg4", "-b:v", "2097152", "-b:a", "48000", "-ac", "2", "-ar", "22050", filePath};
 
         //믹싱에서 바뀌는 파일 미리 생성
         //command_mp4_change = new String[]{"-ss", "" + startMs / 1000, "-y", "-i", ori_Path, "-t", "" + (endMs - startMs) / 1000, "-vcodec", "mpeg4", "-b:v", "2097152", "-b:a", "48000", "-ac", "2", "-ar", "22050", filePath_mp4_change};
@@ -254,9 +254,9 @@ public class videoTrimmer extends AppCompatActivity {
         execffmpegBinary_mp4(command);
         //execffmpegBinary_mp4(command_mp4_change);
 
-        command_mp3 = new String[]{"-ss", "" + startMs / 1000, "-y", "-i", String.valueOf(uri_mp3), "-t", "" + (endMs - startMs) / 1000,"-ac","1", filePath_mp3};
+        command_mp3 = new String[]{"-ss", "" + startMs / 1000, "-y", "-i", String.valueOf(uri_mp3), "-t", "" + (endMs-startMs) / 1000,"-ac","1", filePath_mp3};
         //믹싱에서 바뀌는 목소리 파일 미리 생성
-        command_mp3_change = new String[]{"-ss", "" + startMs / 1000, "-y", "-i", String.valueOf(uri_mp3), "-t", "" + (endMs - startMs) / 1000,"-ac","1", filePath_mp3_change};
+        command_mp3_change = new String[]{"-ss", "" + startMs / 1000, "-y", "-i", String.valueOf(uri_mp3), "-t", "" + (endMs-startMs) / 1000,"-ac","1", filePath_mp3_change};
 
         // command_mp3 = new String[]{"-i", String.valueOf(uri_mp3), "-map","0","-c:v","copy","-af","aecho=0.6:0.5:1000:0.5", filePath_mp3};
